REF move NonlinearDeltaCov to stats._delta_method
